Remove initialPath
This commit also changes the args that are passed to the dispatch
callback. Instead of passing the Transition object, the signature
is now callback(error, abortReason). Of course, both args are null/
undefined if there was neither an error or the transition was not
aborted.

This is going to give us a super-clean API for server-side rendering.