Use createClass
It's annoying to scroll to the bottom of the file just to see propTypes
and contextTypes. By convention, I usually keep this info at the top of
my class definition, which makes it easier for me to scan. Using
createClass lets us keep it up top without resorting to ES7 static properties.